+++ src/common/xml/sorting-rules.ts
@@ -35,6 +35,11 @@ export const DEFAULT_SORTING_RULES: SortingRule[] = [
   {
     filePattern: "labels-meta.xml",
     priorityKeys: ["fullName"]
+  },
+  {
+    filePattern: "globalValueSet-meta.xml",
+    priorityKeys: ["fullName"],
+    unsortedArrays: ["customValue"]
   }
 ];